requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
semgrep==1.52.0
PyGithub==1.59.1
python-gitlab==4.2.0
pyyaml==6.0.1
packaging==23.2
python-dateutil==2.8.2
colorama==0.4.6
rich==13.7.0
typer==0.9.0
aiohttp==3.9.1
asyncio==3.4.3
gitpython==3.1.40